Set card 1 as default for ALSA
Currently, the kernel detects the HDMI soundcard first, so it assigns it
card_0, and then it detects the Conexant CX8070, so it's assigned card_1

By default, ALSA takes card_0 as default. So, we make /etc/asound.conf
with the following configuration to use card_1 as default.

See: https://www.alsa-project.org/wiki/Setting_the_default_device